HMI: return  to old variant of key scanner HMI: fix error with bad time drawing Refactoring of scripts

+++ scripts/getaddrnew.sh
@@ -2,9 +2,13 @@
 source /tmp/about/setenv.sh
 home1=/rw/mx00
 cd $home1/configs
+rm addr.cfg.md5
 sftp $UPDATE@$SFTP:dev/$ID/configs/addr.cfg.md5 $home1/configs > /dev/null 2>&1
 md5sum -c addr.cfg.md5 > /dev/null 2>&1
-if [ $? -ne 0 ] 
+stat $home1/configs/addr.cfg.md5 > /dev/null 2>&1
+if [ $? -eq 0 ]
+  then
+ if [ $? -ne 0 ] 
   then
   rm $home1/configs/addr.cfg > /dev/null 2>&1
   sftp $UPDATE@$SFTP:dev/$ID/configs/addr.cfg $home1/configs > /dev/null 2>&1
@@ -20,5 +24,6 @@ if [ $? -ne 0 ]
     echo "put $home1/cfgupdate.log dev/$ID/logs" | sftp $UPDATE@$SFTP 2>&1 > /dev/null
     reboot
   fi
+ fi
 fi
